Create routes protector function
The `routesProtector` function handles protecting a route
process when it is called inside the `load` function of the
route that should be protected.

The function looks for user's session (that should be created
by Auth.js) to check if the user is signed in to let him continue,
If the user is not signed, the function redirects him to "sign-in"
page.