Fix a bug in `PromptField` component
The bug occurs when a new chat is created, a new chat id comes
from the server and the URL changes to contain the new chat id,
The `chatId` prop of `PromptField` component will remain in its
first state and never change even if the parant component that
passes `chatId` prop to `PromptField` component had the value
of the prop changed.

To fix this problem make `PromptSubmitHandler` function create
the logic of the function that handles submitting the prompts
inside `$derived.by` rune so svelte can track the changes of
whatever dependencies `PromptSubmitHandler` function
use to do its work (`chatId` prop in our case) and rerun it
whenever the dependencies change.