Create dataset page state file for a single dataset
Create `src/stores/datasetPage.svelte.ts` file and inside it
define the state of a single dataset that is supposed to be
opened in `routes/datasets/[datasetId]` page.

Create `src/routes/datasets/[datasetId]/+page.server.ts` route
handler to fetch the full data of the dataset.

Add `DatasetFullData` type into `src/datasets.d.ts` to represent the
full data of the dataset that is supposed to be opened in
`routes/datasets/[datasetId]` page.